% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dir.R
\name{lollipop}
\alias{lollipop}
\title{Graphing function to make lollipop bar graphs}
\usage{
lollipop(df, y_var, x_var, colorme)
}
\arguments{
\item{df}{dataframe source to graph}

\item{y_var}{variable for Y}

\item{x_var}{variable for X}

\item{colorme}{color to make the bars}
}
\description{
Graphing function to make lollipop bar graphs
}
