% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCOE-package.R
\docType{package}
\name{MCOE-package}
\alias{MCOE}
\alias{MCOE-package}
\title{MCOE: Creates New Folders and Loads Standard Libraries for Monterey County Office of Education}
\description{
Basic Setup for other Projects in R for Monterey County Office of Education
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dobrowski/MCOE}
  \item Report bugs at \url{https://github.com/dobrowski/MCOE/issues}
}

}
\author{
\strong{Maintainer}: David Dobrowski \email{ddobrowski@montereycoe.org}

}
\keyword{internal}
