% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dir.R
\name{mcoe_logo}
\alias{mcoe_logo}
\title{Adds the MCOE logo}
\usage{
mcoe_logo()
}
\value{
`mcoe_logo` an object of class `magick-image` from [magick::image_read()] which is an image file of the MCOE logo
}
\description{
Adds the MCOE logo
}
\examples{
mcoe_logo()
}
