% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dir.R
\name{mcoe_sql_con}
\alias{mcoe_sql_con}
\title{Connects to the MCOE 'SQL' tables}
\usage{
mcoe_sql_con()
}
\value{
a database connection in the form of [DBI::dbConnect]
}
\description{
Connects to the MCOE 'SQL' tables
}
\examples{
try(mcoe_sql_con())
}
