% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dir.R
\docType{data}
\name{mcoe_theme}
\alias{mcoe_theme}
\title{Modifies ggplots to use a common theme}
\format{
An object of class \code{list} of length 4.
}
\usage{
mcoe_theme
}
\description{
Modifies ggplots to use a common theme
}
\keyword{datasets}
