% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mry.dist}
\alias{mry.dist}
\title{School District Directory}
\format{
## `mry.dist`
A data frame with 256 rows and 50 columns:
\describe{
  \item{District}{District name}
  \item{School}{School name}
  \item{CDScode}{County, District, School code}
  ...
}
}
\source{
<https://www.cde.ca.gov/ds/si/ds/pubschls.asp>
}
\usage{
mry.dist
}
\description{
A subset of data from California Department of Education identifying schools
}
\keyword{datasets}
