% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dir.R
\name{setup_dir}
\alias{setup_dir}
\title{MCOE
Set up initial project directory}
\usage{
setup_dir()
}
\value{
Two new system directories are made.
}
\description{
MCOE
Set up initial project directory
}
\examples{
setup_dir()
}
