% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dir.R
\name{\%notin\%}
\alias{\%notin\%}
\title{MCOE
The inverse of _in_ where the list is excluded}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{Item or list to look at}

\item{table}{List to make sure it is not part of}
}
\value{
`%notin%` returns a 'list' of class `logical` indicating absence of passed items
}
\description{
MCOE
The inverse of _in_ where the list is excluded
}
\examples{

fruit <- c("apple", "banana", "pear", "pineapple")
"kiwi" \%notin\% fruit
}
