% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCOE-package.R
\docType{package}
\name{MCOE-package}
\alias{MCOE}
\alias{MCOE-package}
\title{MCOE: Creates New Folders and Loads Standard Practices for Monterey County Office of Education}
\description{
Basic Setup for Projects in R for Monterey County Office of Education. It contains functions often used in the analysis of education data in the county office including seeing if an item is not in a list, rounding in the manner the general public expects, including logos for districts, switching between district names and their county-district-school codes, accessing the local 'SQL' table and making thematically consistent graphs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dobrowski/MCOE}
  \item Report bugs at \url{https://github.com/dobrowski/MCOE/issues}
}

}
\author{
\strong{Maintainer}: David Dobrowski \email{ddobrowski@montereycoe.org} [copyright holder]

}
\keyword{internal}
