% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dir.R
\name{mcoe_d_logo}
\alias{mcoe_d_logo}
\title{Given a CDS code, adds the district logo}
\usage{
mcoe_d_logo(cds.number)
}
\arguments{
\item{cds.number}{County-District-School code from California Department of Education}
}
\value{
`mcoe_d_logo` an object of class `magick-image` from [magick::image_read()] which is an image file of the district logo
}
\description{
Given a CDS code, adds the district logo
}
\examples{
try(mcoe_d_logo("27659870000000"))
}
