% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dir.R
\name{mcoe_logo_location}
\alias{mcoe_logo_location}
\title{Given a CDS code it provides the logo}
\usage{
mcoe_logo_location(cds.number)
}
\arguments{
\item{cds.number}{County-District-School code from California Department of Education}
}
\value{
an image file of the logo
}
\description{
Given a CDS code it provides the logo
}
\examples{
mcoe_logo_location("27661590000000")
}
