% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dir.R
\name{round2}
\alias{round2}
\title{Rounding the way most people do it with .5 roudning up}
\usage{
round2(x, digits = 2)
}
\arguments{
\item{x}{number to be rounded}

\item{digits}{how many digits to round to}
}
\value{
`round2` returns a truncated `numeric` with decimal places equal to `digits` argument
}
\description{
Rounding the way most people do it with .5 roudning up
}
\examples{
x <- 1.2345
round2(x, digits = 2)
}
